Fixed over increment of actuator angle and changed PWM frequency
Added flags or 'semaphores' to the actuator movement so the angles cannot be incremented again until the actuator has finished moving to the last position. Flag also released if the angle is within 5 degrees of range limit to prevent locking due to potentiometer movement.

Changed constants within the constrain() function calls to the defined constants.

Increased PWM frequency to 500Hz in stead of 50Hz to allow smoother movement of linear actuators.